+++ public/usage-examples/geometry/point_on_line-1-example-oop.cs
@@ -1,6 +1,6 @@
 using SplashKitSDK;
 
-namespace PointOnLine
+namespace PointOnLineExample
 {
     public class Program
     {
