+++ public/usage-examples/geometry/point_on_line-1-example-top-level.cs
