+++ public/usage-examples/geometry/point_on_line-1-example.cpp
@@ -24,5 +24,6 @@ int main()
         process_events();
     }
     close_window(window);
+    
     return 0;
 }
