--- public/usage-examples/geometry/point_on_line/point_on_line-2-volume-slide.txt
@@ -1,3 +1,0 @@
-#### Point On Line
-
-The following code shows an example of using [Point On Line](/api/geometry#point-on-line) to show a simple use case for point on line in this case as a volume slider
